+++ Breakout/Brick.h
@@ -8,16 +8,16 @@ public:
 	~Brick();
 	void Draw(int raw);
 
-	int getXPosition() { return ActualPosition.x; }
-	int getYPosition() { return ActualPosition.y; }
-	int getWidth() { return Dimensions.Width; }
-	int getHeight() { return Dimensions.Height; }
+	float getXPosition() { return ActualPosition.x; }
+	float getYPosition() { return ActualPosition.y; }
+	float getWidth() { return Dimensions.Width; }
+	float getHeight() { return Dimensions.Height; }
 	bool getActive() { return Active; }
 	bool getTraitor() { return Traitor; }
 	void setActive(bool Change) { Active = Change; }
 	void setTraitor(bool Change) { Traitor = Change; }
 	bool CheckCollition(class Ball* Ball);
-	void setYPosition(int Y) {ActualPosition.y = Y; }
+	void setYPosition(float Y) {ActualPosition.y = Y; }
 
 private:
 	Position ActualPosition;
