+++ Breakout/Ray.h
@@ -11,8 +11,8 @@ public:
 	Position* getPosition() { return &ActualPosition; }
 	Dimension* getDimensions() { return &Dimensions; }
 	void Move(float DeltaTime) { ActualPosition.y -= Velocity; }
-	bool CheckCollition(class Brick* ActualBrick, int DeltaTime, char Power, int WindowHeight);
-	void SetData(int X, int Y, bool _Active);
+	bool CheckCollition(class Brick* ActualBrick, float DeltaTime, char Power, int WindowHeight);
+	void SetData(float X, float Y, bool _Active);
 	bool getActive() { return Active; }
 	
 
@@ -22,5 +22,5 @@ private:
 	int Velocity;
 	bool Active;
 
-	int NYPosition;
+	float NYPosition;
 };
